Add compile-time checks for libandroid_runtime
Before this CL, libandroid_runtime didn't process thread-safety
annotations and didn't enforce the -Wextra checks.

Add those here, along with the required fixes.

Bug: 271455682
Test: m libandroid_runtime

+++ core/jni/com_android_internal_os_LongArrayMultiStateCounter.cpp
@@ -288,7 +288,7 @@ static jboolean native_combineValues_LongArrayContainer(JNIEnv *env, jobject sel
     }
 
     bool nonZero = false;
-    for (int i = 0; i < vector->size(); i++) {
+    for (size_t i = 0; i < vector->size(); i++) {
         jint index = scopedIndexMap[i];
         if (index < 0 || index >= size) {
             jniThrowExceptionFmt(env, "java/lang/IndexOutOfBoundsException",
